lib: fix too-conservative integer overflow check when appending raw name
It looks as if this change should have been included in
25ec4f1b290fcfee9e8299c6ec17d0ace800136c.

+++ expat/lib/xmlparse.c
@@ -3063,7 +3063,7 @@ storeRawNames(XML_Parser parser) {
     */
     rawNameLen = ROUND_UP(tag->rawNameLength, sizeof(XML_Char));
     /* Detect and prevent integer overflow. */
-    if (rawNameLen > (size_t)INT_MAX - nameLen)
+    if (rawNameLen > SIZE_MAX - nameLen)
       return XML_FALSE;
     bufSize = nameLen + rawNameLen;
     if (bufSize > (size_t)(tag->bufEnd - tag->buf.raw)) {
